refactor: address PR review comments on SQLite store
- Add doc comment explaining :memory: DSN on NewMemoryStore
- Enforce single file-backed store per process via mutex guard
- Move migrations to sqlite_migrations.go
- Promote scanner to named RowScanner type in store.go
- Reorganize: store.go (types+interface), sqlite_store.go (impl),
  sqlite_migrations.go (schema)
